MyApps-13: Move local web implementation, api and dependency injection to submodules.
