+++ app/src/main/java/xyz/torquato/myapps/di/data/DataModule.kt
@@ -5,10 +5,8 @@ import dagger.Module
 import dagger.hilt.InstallIn
 import dagger.hilt.components.SingletonComponent
 import xyz.torquato.myapps.api.math.IMathRepository
-import xyz.torquato.myapps.api.web.ILocalWebRepository
 import xyz.torquato.myapps.api.web.cache.IWebCacheRepository
 import xyz.torquato.myapps.data.math.MathRepository
-import xyz.torquato.myapps.data.web.LocalWebRepository
 import xyz.torquato.myapps.data.web.cache.WebCacheRepository
 import javax.inject.Singleton
 
@@ -28,8 +26,6 @@ abstract class RepositoryModule {
 
 
 
-    @Binds
-    @Singleton
-    abstract fun bindLocalWebImpl(impl: LocalWebRepository): ILocalWebRepository
+
 
 }
