+++ app/src/main/java/xyz/torquato/myapps/di/domain/DomainWebModule.kt
@@ -4,7 +4,6 @@ import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
 import dagger.hilt.components.SingletonComponent
-import xyz.torquato.myapps.api.web.ILocalWebRepository
 import xyz.torquato.myapps.api.web.cache.IWebCacheRepository
 import xyz.torquato.myapps.domain.web.GetSelectedBookUseCase
 import xyz.torquato.myapps.domain.web.SetSelectedBookUseCase
@@ -12,6 +11,7 @@ import xyz.torquato.myapps.domain.web.content.GetBooksUseCase
 import xyz.torquato.myapps.domain.web.content.GetCacheBooksUseCase
 import xyz.torquato.myapps.domain.web.pagging.GetNextPageUseCase
 import xyz.torquato.myaps.domain.api.pagging.IPagingRepository
+import xyz.torquato.myaps.domain.api.web.ILocalWebRepository
 import xyz.torquato.myaps.domain.api.web.IQueryRepository
 import xyz.torquato.myaps.domain.api.web.IWebRepository
 import javax.inject.Singleton
