+++ app/src/main/java/xyz/torquato/myapps/domain/web/GetSelectedBookUseCase.kt
@@ -1,6 +1,6 @@
 package xyz.torquato.myapps.domain.web
 
-import xyz.torquato.myapps.api.web.ILocalWebRepository
+import xyz.torquato.myaps.domain.api.web.ILocalWebRepository
 
 class GetSelectedBookUseCase(
     private val repository: ILocalWebRepository
