+++ data/di/src/main/java/xyz/torquato/myapps/data/di/RepositoryImpl.kt
@@ -7,11 +7,13 @@ import dagger.hilt.components.SingletonComponent
 import xyz.torquato.myapps.data.impl.cypher.CypherRepository
 import xyz.torquato.myapps.data.impl.paging.PagingRepository
 import xyz.torquato.myapps.data.impl.waves.SoundRepository
+import xyz.torquato.myapps.data.impl.web.LocalWebRepository
 import xyz.torquato.myapps.data.impl.web.QueryRepository
 import xyz.torquato.myapps.data.impl.web.WebRepository
 import xyz.torquato.myaps.domain.api.encrypt.ICypherRepository
 import xyz.torquato.myaps.domain.api.pagging.IPagingRepository
 import xyz.torquato.myaps.domain.api.sound.ISoundRepository
+import xyz.torquato.myaps.domain.api.web.ILocalWebRepository
 import xyz.torquato.myaps.domain.api.web.IQueryRepository
 import xyz.torquato.myaps.domain.api.web.IWebRepository
 import javax.inject.Singleton
@@ -38,5 +40,9 @@ abstract class RepositoryImpl {
     @Binds
     @Singleton
     abstract fun bindQueryImpl(impl: QueryRepository): IQueryRepository
+
+    @Binds
+    @Singleton
+    abstract fun bindLocalWebImpl(impl: LocalWebRepository): ILocalWebRepository
 
 }
