+++ data/impl/src/main/java/xyz/torquato/myapps/data/impl/web/LocalWebDataSource.kt
@@ -1,4 +1,4 @@
-package xyz.torquato.myapps.data.web
+package xyz.torquato.myapps.data.impl.web
 
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.MutableSharedFlow
